requirements.txt
pysam>=0.19.0
numpy>=1.21.0
scipy>=1.7.0
pyyaml>=6.0

# Development dependencies
pytest>=7.0.0
pytest-cov>=4.0.0